+++ src/store/index.js
@@ -1,5 +1,4 @@
 import { configureStore } from "@reduxjs/toolkit";
-//reducer
 import user from './slices/user';
 import activities from "./slices/activities";
